Moved resource type files to the store folder
