+++ src/store/mod.rs
@@ -1,27 +1,14 @@
 use super::*;
 use getset::CopyGetters;
 
-mod builder;
-mod many_to_many;
-mod one_to_many;
-mod relationships_options;
-mod relationships_options_builder;
-mod resource_type_configuration;
-
-pub use builder::CibouletteStoreBuilder;
-pub use relationships_options::{
-    CibouletteRelationshipManyToManyOption, CibouletteRelationshipOneToManyOption,
-    CibouletteRelationshipOption,
-};
-pub use relationships_options_builder::{
-    CibouletteRelationshipManyToManyOptionBuilder, CibouletteRelationshipOneToManyOptionBuilder,
-    CibouletteRelationshipOptionBuilder,
-};
-pub use resource_type_configuration::{
-    CiboulettePaginationEncoding, CibouletteResourceTypeConfiguration,
-    CibouletteResourceTypePaginationConfiguration,
-    CibouletteResourceTypePaginationConfigurationBuilder,
-};
+pub mod builder;
+pub mod many_to_many;
+pub mod one_to_many;
+pub mod relationships_options;
+pub mod relationships_options_builder;
+pub mod resource_type;
+pub mod resource_type_builder;
+pub mod resource_type_configuration;
 
 #[cfg(test)]
 mod tests;
