+++ src/store/resource_type_builder.rs
